requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi     #==2020.6.20        # via -r requirements.in, requests
cffi     #==1.14.3              # via cryptography
chardet     ==4.0.0           # via -r requirements.in, requests
click     #==7.1.2              # via pip-tools
colorama     #==0.4.3           # via -r requirements.in, crayons
configparser     #==5.0.0       # via -r requirements.in, webdriver-manager
crayons     #==0.4.0            # via -r requirements.in, webdriver-manager
execnet     #==1.7.1            # via -r requirements.in, pytest-xdist
glob2     #==0.7                # via -r requirements.in, pytest-bdd
idna     #==2.10                # via -r requirements.in, requests
importlib-metadata     #==2.0.0  # via -r requirements.in
iniconfig     #==1.0.1          # via -r requirements.in, pytest
mako     #==1.1.3               # via -r requirements.in, pytest-bdd
markupsafe     #==1.1.1         # via -r requirements.in, mako
more-itertools     #==8.5.0     # via -r requirements.in
packaging     #==20.4           # via -r requirements.in, pytest
parse-type     #==0.5.2         # via -r requirements.in, pytest-bdd
parse     #==1.18.0             # via -r requirements.in, parse-type, pytest-bdd
pip-tools     #==5.3.1          # via -r requirements.in
pluggy     #==0.13.1            # via -r requirements.in, allure-python-commons, pytest
psutil     #==5.8.0             # via browserstack-local
py     #==1.9.0                 # via -r requirements.in, pytest, pytest-bdd, pytest-forked
pycparser     #==2.20           # via cffi
pyparsing     #==2.4.7          # via -r requirements.in, packaging
pytest-bdd     #==6.0.1         # via -r requirements.in
pytest-forked     #==1.3.0      # via -r requirements.in, pytest-xdist
pytest-html     #==2.1.1        # via -r requirements.in
pytest-metadata     #==1.10.0   # via pytest-html
pytest-repeat     #==0.9.1      # via -r requirements.in
pytest-xdist     #==2.1.0       # via -r requirements.in
pytest     #==7.1.2             # via -r requirements.in, allure-pytest, pytest-bdd, pytest-forked, pytest-html, pytest-metadata, pytest-repeat, pytest-xdist
pytz     #==2021.1              # via -r requirements.in
requests     #==2.24.0          # via -r requirements.in, webdriver-manager
rstr     #==2.2.6               # via -r requirements.in
selenium==4.1.2                 # via -r requirements.in
Appium-Python-Client  #~=2.7.1     # via -r requirements.in
six     #==1.15.0               # via -r requirements.in, allure-pytest, allure-python-commons, cryptography, packaging, parse-type, pip-tools, pytest-bdd
toml     #==0.10.2              # via -r requirements.in, pytest
urllib3==1.26.8          # via -r requirements.in, requests, selenium
webdriver-manager     #==3.2.2  # via -r requirements.in
wheel     #==0.35.1             # via -r requirements.in
zipp     #==3.4.0               # via -r requirements.in, importlib-metadata
#browsermob-proxy     #==0.8.0   #via -r requirements.in, import browsermobproxy
# The following packages are considered to be unsafe in a requirements file:
# pip

setuptools  #~=60.5.0

